auto

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: adaly <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/05/16 06:37:13 by adaly             #+#    #+#             */
-/*   Updated: 2017/05/27 01:04:46 by adaly            ###   ########.fr       */
+/*   Updated: 2017/05/27 01:12:58 by adaly            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,9 +69,9 @@ void		ft_printf_int(t_pfconv *current, va_list args)
 		if (current->precision == -2)
 			current->precision = va_arg(args, int);
 		if (current->import_type == 'i')
-			ft_eval_num(current, va_arg(args, int));
+			ft_eval_num(current, (long long)va_arg(args, int));
 		if (current->import_type == 'l')
-			ft_eval_num(current, va_arg(args, long int));
+			ft_eval_num(current, (long long)va_arg(args, long int));
 		if (current->import_type == 'm')
 			ft_eval_num(current, va_arg(args, long long int));
 		if (current->import_type == 'u')
